Make indexing more efficient
Rather than keeping track of these on a character by character basis (i.e.
mapping every single position to an offset), just keep a running offset-per
line. This happens to be easy to calculate: offset = sum of line lengths of
previous lines. Given that, looking up the offset for a position becomes a
matter of adding the column to the offset.